+++ app/channels/chatroom_channel.rb
@@ -3,8 +3,4 @@ class ChatroomChannel < ApplicationCable::Channel
     chatroom = Chatroom.find(params[:id])
     stream_for chatroom
   end
-
-  def unsubscribed
-    # Any cleanup needed when channel is unsubscribed
-  end
 end
