+++ app/javascript/controllers/chatroom_subscription_controller.js
@@ -1,19 +1,41 @@
-// app/javascript/controllers/chatroom_subscription_controller.js
 import { Controller } from "@hotwired/stimulus"
 import { createConsumer } from "@rails/actioncable"
 
 export default class extends Controller {
-  static values = { chatroomId: Number }
+  static values = {
+    chatroomId: Number,
+    userId: Number
+  }
   static targets = ["messages"]
 
-  constructor(...args) {
-    super(...args)
-    this.channel = null
+  connect() {
+    console.log('hello');
+    this.channel = createConsumer().subscriptions.create(
+      { channel: "ChatroomChannel", id: this.chatroomIdValue },
+      { received: res => {
+        const data = JSON.parse(res)
+        console.log(data)
+        this.#insertMessageAndScrollDown(data.elementToAppend)
+        console.log(data.authorId)
+        console.log(this.userIdValue)
+        if (Number(data.authorId) === Number(this.userIdValue)){
+          this.messagesTarget.lastElementChild.classList.remove("other-message")
+          this.messagesTarget.lastElementChild.classList.add("my-message")
+        }
+      }}
+    )
+    console.log(`Subscribed to the chatroom with the id ${this.chatroomIdValue}.`)
   }
 
-  connect() {
-    console.log(`Subscribing to Chatroom with the id ${this.chatroomIdValue}`)
-    this.subscribe()
+
+  #insertMessageAndScrollDown(data) {
+    this.messagesTarget.insertAdjacentHTML("beforeend", data)
+    this.messagesTarget.scrollTo(0, this.messagesTarget.scrollHeight)
+  }
+
+  resetForm(event) {
+    console.log('oui');
+    event.target.reset()
   }
 
   disconnect() {
@@ -21,53 +43,4 @@ export default class extends Controller {
     this.channel.unsubscribe()
   }
 
-  received(data) {
-    this.insertMessage(data)
-    this.insertMessageAndScrollDown(data)
-    this.resetForm()
-  }
-
-  insertMessage(data) {
-    this.messagesTarget.insertAdjacentHTML("beforeend", data)
-  }
-
-  insertMessageAndScrollDown(data) {
-    this.messagesTarget.scrollTo(0, this.messagesTarget.scrollHeight)
-  }
-
-  resetForm() {
-    const form = this.element.querySelector("form")
-    if (form) {
-      form.reset()
-    }
-  }
-
-  subscribe() {
-    const consumer = createConsumer()
-    this.channel = consumer.subscriptions.create(
-      { channel: "ChatroomChannel", id: this.chatroomIdValue },
-      {
-        received: this.received.bind(this)
-      }
-    )
-  }
-
-  deleteMessage(event) {
-    const messageElement = event.target.closest(".message")
-    const messageId = messageElement.dataset.messageId
-
-    fetch(`/events/${this.chatroomIdValue}/chatrooms/${this.chatroomIdValue}/messages/${messageId}`, {
-      method: "DELETE"
-    })
-      .then(response => {
-        if (response.ok) {
-          messageElement.remove()
-        } else {
-          console.error("Failed to delete message")
-        }
-      })
-      .catch(error => {
-        console.error("Error occurred while deleting message:", error)
-      })
-  }
 }
